chore: move all state structs to
