+++ listing-rewards/program/src/assertions.rs
@@ -2,8 +2,9 @@ use anchor_lang::prelude::*;
 use mpl_auction_house::state::AuctionHouse;
 
 use crate::{
-    errors::ListingRewardsError, reward_center::RewardCenter,
-    rewardable_collection::RewardableCollection, sell::Listing, MetadataAccount,
+    errors::ListingRewardsError,
+    state::{Listing, RewardCenter, RewardableCollection},
+    MetadataAccount,
 };
 
 pub fn assert_listing_reward_redemption_eligibility(
