+++ listing-rewards/program/src/offers/mod.rs
@@ -1,1 +1,2 @@
+pub mod close;
 pub mod create;
