+++ listing-rewards/program/src/redeem_rewards/mod.rs
@@ -8,9 +8,7 @@ use crate::{
     assertions::assert_listing_reward_redemption_eligibility,
     constants::{LISTING, REWARDABLE_COLLECTION, REWARD_CENTER},
     errors::ListingRewardsError,
-    reward_center::RewardCenter,
-    rewardable_collection::RewardableCollection,
-    sell::Listing,
+    state::{Listing, RewardCenter, RewardableCollection},
     MetadataAccount,
 };
